pyproject.toml
[tool.poetry]
name = "cricgenius"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
pandas = "^2.2.3"
requests = "^2.32.3"
tqdm = "^4.67.0"
torch = "^2.5.1"
captum = "^0.7.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.3"
httpx = "^0.28.0"
tkcalendar = "^1.6.1"
scikit-learn = "^1.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
